+++ src/activity/Activity.js
@@ -958,7 +958,7 @@ export default function Activity(Behaviour, activityDef, context) {
 
       popFormattingStart(routingKey, message);
 
-      logger.debug(`<${id}> format ${fields.routingKey} message content`);
+      logger.debug(`<${id}> format ${fields.routingKey} message content with formatter ${routingKey}`);
 
       formattedContent = {
         ...formattedContent,